backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
idna==3.4
nba-api==1.1.14
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
